Address PR feedback for #812
