+++ api_service/api/routers/oauth_sessions.py
@@ -7,6 +7,7 @@ from sqlalchemy.future import select
 from api_service.db.base import get_async_session
 from api_service.auth_providers import get_current_user
 from api_service.api.schemas_oauth_sessions import CreateOAuthSessionRequest, OAuthSessionResponse
+from api_service.services.auth_profile_service import sync_auth_profile_manager
 from api_service.db.models import ManagedAgentOAuthSession, OAuthSessionStatus, User, ManagedAgentAuthProfile, ManagedAgentAuthMode, ManagedAgentRateLimitPolicy
 
 router = APIRouter(prefix="/oauth-sessions", tags=["oauth-sessions"])
@@ -145,8 +146,8 @@ async def finalize_oauth_session(
     if not session_obj:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session not found")
         
-    if session_obj.status in [OAuthSessionStatus.CANCELLED, OAuthSessionStatus.EXPIRED, OAuthSessionStatus.FAILED]:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Cannot finalize a terminal session")
+    if session_obj.status not in [OAuthSessionStatus.AWAITING_USER, OAuthSessionStatus.VERIFYING]:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"Cannot finalize session in {session_obj.status.name} state")
         
     session_obj.status = OAuthSessionStatus.SUCCEEDED
     session_obj.completed_at = datetime.now(timezone.utc)
@@ -164,31 +165,40 @@ async def finalize_oauth_session(
     try:
         policy_enum = ManagedAgentRateLimitPolicy(policy_str)
     except ValueError:
-        policy_enum = ManagedAgentRateLimitPolicy.BACKOFF
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Unsupported rate_limit_policy: {policy_str}"
+        )
+
+    if existing_profile and existing_profile.owner_user_id is not None and str(existing_profile.owner_user_id) != str(current_user.id):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this profile")
+
+    profile_data = {
+        "runtime_id": session_obj.runtime_id,
+        "auth_mode": ManagedAgentAuthMode.OAUTH,
+        "volume_ref": session_obj.volume_ref,
+        "volume_mount_path": session_obj.volume_mount_path,
+        "account_label": session_obj.account_label,
+        "max_parallel_runs": metadata.get("max_parallel_runs", 1),
+        "cooldown_after_429_seconds": metadata.get("cooldown_after_429_seconds", 300),
+        "rate_limit_policy": policy_enum,
+        "enabled": True,
+    }
 
     if existing_profile:
-        existing_profile.runtime_id = session_obj.runtime_id
-        existing_profile.auth_mode = ManagedAgentAuthMode.OAUTH
-        existing_profile.volume_ref = session_obj.volume_ref
-        existing_profile.account_label = session_obj.account_label
-        existing_profile.max_parallel_runs = metadata.get("max_parallel_runs", 1)
-        existing_profile.cooldown_after_429_seconds = metadata.get("cooldown_after_429_seconds", 300)
-        existing_profile.rate_limit_policy = policy_enum
-        existing_profile.enabled = True
+        for key, value in profile_data.items():
+            setattr(existing_profile, key, value)
     else:
         new_profile = ManagedAgentAuthProfile(
             profile_id=session_obj.profile_id,
-            runtime_id=session_obj.runtime_id,
-            auth_mode=ManagedAgentAuthMode.OAUTH,
-            volume_ref=session_obj.volume_ref,
-            account_label=session_obj.account_label,
-            max_parallel_runs=metadata.get("max_parallel_runs", 1),
-            cooldown_after_429_seconds=metadata.get("cooldown_after_429_seconds", 300),
-            rate_limit_policy=policy_enum,
-            enabled=True
+            owner_user_id=current_user.id,
+            **profile_data
         )
         db.add(new_profile)
 
     await db.commit()
     
+    await sync_auth_profile_manager(session=db, runtime_id=session_obj.runtime_id)
+    
     return {"status": "succeeded"}
+
